ref: transactions
this refactor includes changes to how transaction input data is handled.
prevously input data was handled in the Transactions/index.js, which
required prop-drilling and risked sending incorrect inputs.
now, each transaction's input is kept within the component that is
sending the transaction, and conversions are handled in sagas, where the
client cannot mutate

* move transaction logic into sagas
* add state to each transaction side panel
* add handlers for transaction input changes
* extract signedMsg saga